Do not use OpenNI/NiTE listeners to read frames
OpenNI/NiTE listeners were used before in order to read depth, color, and
usertracker frames. However, they are not required because own OpenNI*Instance
classes are designed to behave as a driver. So, only one instance of each class
is required in runtime. Because of this, listeners are not needed and readFrame
methods of OpenNI are used instead.

In addition, a new OpenNIUserTrackerInstance has been implemented in order to
read user frame and skeleton frame from the same depth frame.