Simplify PlaybackControl
PlaybackControl has been simplified. A PlaybackControl reads frames from
several instances and notify viewers about it. Thats all!. No waits for
viewers to render frames on screen. The viewer is in charge of discarding
a given frame if that frame is received out of time. In other words, if
a viewer is notified about new frames while it's still performing a task,
the notification will be received out of time and the received frame will
not be available for read because it will have been overwrited by a new
one.