Improve OpenNI support so that HuDaAct can create OpenNI Instances
OpenNI needed to be improved in order to be used from HuDaAct dataset.
Because of this, OpenNIRuntime is OpenNIDevice now. An OpenNIDevice is
not singleton, so multiple devices can be used at the same time. Morever,
each OpenNIInstance is composed of an OpenNIDevice now. This device can be
created by the own instance or given, so that it can be shared by other
instances.

Because of OpenNI are now part of CoreLib, config folder of the
PrivacyFilter project has been moved to KinectProject folder. This
config folder includes the OpenNI settings needed by PrivacyFilter and
DatasetBrowser projects.